agregando el controller y service guardar pago

+++ src/main/java/com/app/bematdid/dto/PagoDTO.java
@@ -24,5 +24,5 @@ public class PagoDTO implements Serializable {
     private Date fecha;
     private String comentario;
     private String tipoPago;
-    private Double monto;
+    private Integer monto;
 }
